requirements.txt
django>=5.0,<6.0
psycopg2-binary
django-q2
django-htmx
django-compressor
requests
gunicorn
whitenoise
ping3
